backend/requirements.txt
asgiref==3.5.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.4
liqpay-sdk-python3==1.0.6
Pillow==9.2.0
PyJWT==2.4.0
pytz==2022.2.1
requests==2.31.0
sqlparse==0.4.2
urllib3==2.0.4
